+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
 - Added spawnContainers API to agent app and agent service
 - Allow configuring port for HTTP instances of AgentServiceApp & SequenceManagerApp from command line.
 - Updated app name's with `esw` prefix according to `apps.prod.json` in `osw-apps` repo.
+- Removed `Struct` script dsl helpers. The idea of removal was sent out with the last release.
 
 ## [ESW v0.2.1] - 2021-01-29